// web/src/api/models/LayoutMonitor.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LayoutMonitor = {
    name: string;
    edid: string;
    port: string;
    width: number;
    height: number;
    refresh_rate: number;
    position_x: number;
    position_y: number;
    primary: boolean;
    /**
     * logical scale factor (1 = 100%, 2 = 200%, 1.5 = 150% fractional)
     */
    scale: number;
};